refactor: 🔨 Split field into fieldname and modifier
