+++ sigma/validators/sigmahq/fieldname.py
@@ -134,96 +134,6 @@ class SigmahqInvalidFieldnameValidator(SigmaDetectionItemValidator):
 
 
 @dataclass
-class SigmahqFieldDuplicateValueIssue(SigmaValidationIssue):
-    description: ClassVar[str] = "Field list value have a dulicate item"
-    severity: ClassVar[SigmaValidationIssueSeverity] = SigmaValidationIssueSeverity.HIGH
-    field: str
-    value: str
-
-
-class SigmahqFieldDuplicateValueValidator(SigmaDetectionItemValidator):
-    """Check uniques value in field list."""
-
-    def validate(self, rule: SigmaRule | SigmaCorrelationRule) -> List[SigmaValidationIssue]:
-        # Only validate SigmaRule (detection rules), not correlation rules
-        if not isinstance(rule, SigmaRule):
-            return []
-        return super().validate(rule)
-
-    def validate_detection_item(
-        self, detection_item: SigmaDetectionItem
-    ) -> List[SigmaValidationIssue]:
-        # Special case where value is case sensitive
-        if (
-            SigmaBase64Modifier in detection_item.modifiers
-            or SigmaBase64OffsetModifier in detection_item.modifiers
-            or SigmaRegularExpressionDotAllFlagModifier in detection_item.modifiers
-            or SigmaRegularExpressionFlagModifier in detection_item.modifiers
-            or SigmaRegularExpressionIgnoreCaseFlagModifier in detection_item.modifiers
-            or SigmaRegularExpressionModifier in detection_item.modifiers
-            or SigmaRegularExpressionMultilineFlagModifier in detection_item.modifiers
-            or SigmaCaseSensitiveModifier in detection_item.modifiers
-        ):
-            value_see = []
-            for v in detection_item.value:
-                if v in value_see:
-                    if detection_item.field is not None:
-                        return [
-                            SigmahqFieldDuplicateValueIssue(
-                                [self.rule], detection_item.field, str(v)
-                            )
-                        ]
-                    else:
-                        return []
-                else:
-                    value_see.append(v)
-            return []
-        else:
-            value_see = []
-            for v in detection_item.value:
-                if str(v).lower() in value_see:
-                    if detection_item.field is not None:
-                        return [
-                            SigmahqFieldDuplicateValueIssue(
-                                [self.rule], detection_item.field, str(v)
-                            )
-                        ]
-                    else:
-                        return []
-                else:
-                    value_see.append(str(v).lower())
-            return []
-
-
-@dataclass
-class SigmahqInvalidAllModifierIssue(SigmaValidationIssue):
-    description: ClassVar[str] = "All modifier without a list of value"
-    severity: ClassVar[SigmaValidationIssueSeverity] = SigmaValidationIssueSeverity.HIGH
-    field: str
-
-
-class SigmahqInvalidAllModifierValidator(SigmaDetectionItemValidator):
-    """Check All modifier used with a single value."""
-
-    def validate(self, rule: SigmaRule | SigmaCorrelationRule) -> List[SigmaValidationIssue]:
-        # Only validate SigmaRule (detection rules), not correlation rules
-        if not isinstance(rule, SigmaRule):
-            return []
-        return super().validate(rule)
-
-    def validate_detection_item(
-        self, detection_item: SigmaDetectionItem
-    ) -> List[SigmaValidationIssue]:
-        if SigmaAllModifier in detection_item.modifiers and len(detection_item.value) < 2:
-            if detection_item.field is not None:
-                return [SigmahqInvalidAllModifierIssue([self.rule], detection_item.field)]
-            else:
-                return []
-        else:
-            return []
-
-
-@dataclass
 class SigmahqFieldUserIssue(SigmaValidationIssue):
     description: ClassVar[str] = "User Field has a Localized name"
     severity: ClassVar[SigmaValidationIssueSeverity] = SigmaValidationIssueSeverity.HIGH
