+++ tests/sigmahq/fieldname/test_sigmahq_field_user_validator.py
@@ -1,6 +1,6 @@
 from sigma.correlations import SigmaCorrelationRule
 from sigma.rule import SigmaRule
-from sigma.validators.sigmahq.field import (
+from sigma.validators.sigmahq.fieldname import (
     SigmahqFieldUserIssue,
     SigmahqFieldUserValidator,
 )
