+++ tests/sigmahq/fieldname/test_sigmahq_fieldname_cast_validator.py
@@ -1,6 +1,6 @@
 from sigma.rule import SigmaRule
 from sigma.correlations import SigmaCorrelationRule
-from sigma.validators.sigmahq.field import (
+from sigma.validators.sigmahq.fieldname import (
     SigmahqFieldnameCastIssue,
     SigmahqFieldnameCastValidator,
 )
