+++ tests/sigmahq/fieldname/test_sigmahq_redundant_field_validator,.py
@@ -1,7 +1,7 @@
 from sigma.rule import SigmaRule
 from sigma.correlations import SigmaCorrelationRule
 
-from sigma.validators.sigmahq.field import (
+from sigma.validators.sigmahq.fieldname import (
     SigmahqRedundantFieldIssue,
     SigmahqRedundantFieldValidator,
 )
