+++ tests/sigmahq/modifier/test_sigmahq_invalid_all_modifier_validator.py
@@ -1,7 +1,7 @@
 from sigma.rule import SigmaRule
 from sigma.correlations import SigmaCorrelationRule
 
-from sigma.validators.sigmahq.field import (
+from sigma.validators.sigmahq.modifier import (
     SigmahqInvalidAllModifierIssue,
     SigmahqInvalidAllModifierValidator,
 )
